Trait libgssapi::context::SecurityContext

pub trait SecurityContext {
    fn wrap(&self, encrypt: bool, msg: &[u8]) -> Result<Buf, Error>;

    fn unwrap(&self, msg: &[u8]) -> Result<Buf, Error>;

    fn info(&self) -> Result<CtxInfo, Error>;

    fn source_name(&self) -> Result<Name, Error>;

    fn target_name(&self) -> Result<Name, Error>;

    fn lifetime(&self) -> Result<Duration, Error>;

    fn mechanism(&self) -> Result<&'static Oid, Error>;

    fn flags(&self) -> Result<CtxFlags, Error>;

    fn local(&self) -> Result<bool, Error>;

    fn open(&self) -> Result<bool, Error>;
}

Required methods

fn wrap(&self, encrypt: bool, msg: &[u8]) -> Result<Buf, Error>

Wrap a message with optional encryption. If encrypt is true then only the other side of the context can read the message. In any case the other side can always verify message integrity.

fn unwrap(&self, msg: &[u8]) -> Result<Buf, Error>

Unwrap a wrapped message, checking it's integrity and decrypting it if necessary.

fn info(&self) -> Result<CtxInfo, Error>

Get all information about a security context in one call

fn source_name(&self) -> Result<Name, Error>

Get the source name of the security context

fn target_name(&self) -> Result<Name, Error>

Get the target name of the security context

fn lifetime(&self) -> Result<Duration, Error>

Get the lifetime of the security context

fn mechanism(&self) -> Result<&'static Oid, Error>

Get the mechanism of the security context

fn flags(&self) -> Result<CtxFlags, Error>

Get the flags of the security context

fn local(&self) -> Result<bool, Error>

Return true if the security context was locally initiated

fn open(&self) -> Result<bool, Error>

Return true if the security context is open

Loading content...